% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmatrix.R
\name{basic_matrix_ops}
\alias{basic_matrix_ops}
\alias{as_mat}
\alias{as_row}
\alias{as_col}
\title{Basic matrix ops}
\usage{
as_mat(x)

as_row(x)

as_col(x)
}
\description{
These will always yield a (dense) \code{base::matrix}.
For \code{as_row()} and \code{as_col()}, that'll be a one-row or one-col matrix, but still a matrix.
}
