% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmatrix.R
\name{cosine_sim}
\alias{cosine_sim}
\alias{vec_cos_sim}
\alias{mat_row_cos_sim}
\alias{mat_col_cos_sim}
\alias{mat_margin_cos_sim}
\title{Cosine similarities}
\usage{
vec_cos_sim(vec_x, vec_y)

mat_row_cos_sim(mat_l, mat_r = mat_l)

mat_col_cos_sim(mat_l, mat_r = mat_l)

mat_margin_cos_sim(mat_l, mat_r = mat_l, margin = 2)
}
\description{
Cosine similarities
}
