% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmatrix.R
\name{lp_normalizers}
\alias{lp_normalizers}
\alias{vec_lpnormalize}
\alias{mat_row_lpnormalize}
\alias{mat_col_lpnormalize}
\alias{mat_margin_lpnormalize}
\title{Lp-normalizers}
\usage{
vec_lpnormalize(vec, p = 1)

mat_row_lpnormalize(mat, p = 1)

mat_col_lpnormalize(mat, p = 1)

mat_margin_lpnormalize(mat, p = 1, margin = 2)
}
\description{
Lp-normalizers
}
