% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmatrix.R
\name{lp_norms}
\alias{lp_norms}
\alias{vec_lpnorm}
\alias{mat_margin_lpnorms}
\alias{mat_col_lpnorms}
\alias{mat_row_lpnorms}
\title{Lp-norms}
\usage{
vec_lpnorm(vec, p = 1)

mat_margin_lpnorms(mat, p = 1, margin = 2)

mat_col_lpnorms(mat, p = 1)

mat_row_lpnorms(mat, p = 1)
}
\description{
Lp-norms
}
