% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{mat_chunk_iixs}
\alias{mat_chunk_iixs}
\title{Matrix chunking indices.}
\usage{
mat_chunk_iixs(
  mat,
  margin,
  chunk_count = NULL,
  chunk_max_length = NULL,
  chunk_max_bytes = NULL,
  cell_bytes = 8
)
}
\description{
Chunking a matrix means preserving at least one of the two dimensions.
In the pathological extreme case the smallest chunk is either a single col or a single row.
No subset of \emph{both} rows and cols are supported.
}
