% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{mat_margin_lpnorms_chunked}
\alias{mat_margin_lpnorms_chunked}
\title{Margin norms computed chunk-wise}
\usage{
mat_margin_lpnorms_chunked(
  mat,
  margin = 2,
  p = 1,
  chunk_count = NULL,
  chunk_max_length = NULL,
  chunk_max_bytes = NULL,
  cell_bytes = 8,
  mc.cores = 1
)
}
\description{
Computes margin norms for a matrix but in chunks.
Notably this allows for parallel computation, which is supported out-of-the-box using \code{parallel::mclapply} and the \code{mc.cores} parameter.
}
