% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murlib.cfmatrix-package.R
\docType{package}
\name{murlib.cfmatrix-package}
\alias{murlib.cfmatrix}
\alias{murlib.cfmatrix-package}
\title{murlib.cfmatrix: Collaborative Filter Matrix Utils}
\description{
Common calculations and transformations used in collaborative filter code. Useful for other matrix-heavy code, too.
}
\author{
\strong{Maintainer}: Murat Tasan \email{mmuurr@gmail.com} (\href{https://orcid.org/0000-0002-8155-2565}{ORCID})

}
\keyword{internal}
