% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmatrix.R
\name{vector_binary_product_ops}
\alias{vector_binary_product_ops}
\alias{vec_inner_prod}
\alias{vec_outer_prod}
\title{Vector binary product ops}
\usage{
vec_inner_prod(vec_x, vec_y = vec_x)

vec_outer_prod(vec_l, vec_r = vec_l)
}
\description{
Vector binary product ops
}
